+++ src/pages/articles/categories/[slug].page.tsx
@@ -2,11 +2,11 @@ import type { ParsedUrlQuery } from "node:querystring";
 
 import type { GetStaticPaths, GetStaticProps, NextPage } from "next";
 
-import { fetchCategories, fetchConfig, fetchTags } from "@/api";
 import type { CategoryProps } from "@/components/pages/articles/categories/Category";
 import { Category } from "@/components/pages/articles/categories/Category";
 import { sentryLogServer } from "@/lib/sentry/logger";
 import ErrorPage from "@/pages/_error/index.page";
+import { fetchCategories, fetchConfig, fetchTags } from "@/repositories";
 import { getArticles, getPickupArticles, getPopularArticles } from "@/services/article";
 import { getCategories, getCategory } from "@/services/category";
 import type { PagePropsOrError } from "@/types";
