+++ src/pages/preview/[id].page.tsx
@@ -1,13 +1,13 @@
 import type { GetServerSideProps, NextPage } from "next";
 import type { Params } from "next/dist/shared/lib/router/utils/route-matcher";
 
-import { fetchConfig, fetchTags } from "@/api";
 import { HtmlHeadNoIndex } from "@/components/functions/meta";
 import { mdx2html } from "@/lib/mdx";
 import { sentryLogServer } from "@/lib/sentry/logger";
 import ErrorPage from "@/pages/_error/index.page";
 import type { ArticleDetailPageProps } from "@/pages/articles/[id].page";
 import { ArticleDetailPage } from "@/pages/articles/[id].page";
+import { fetchConfig, fetchTags } from "@/repositories";
 import { getArticle, getPickupArticles, getPopularArticles } from "@/services/article";
 import { getCategories } from "@/services/category";
 import type { TArticle } from "@/types";
