+++ src/pages/search/index.page.tsx
@@ -1,10 +1,10 @@
 import type { GetStaticProps, NextPage } from "next";
 
-import { fetchConfig, fetchTags } from "@/api";
 import type { SearchProps } from "@/components/pages/Search";
 import { Search } from "@/components/pages/Search";
 import { sentryLogServer } from "@/lib/sentry/logger";
 import ErrorPage from "@/pages/_error/index.page";
+import { fetchConfig, fetchTags } from "@/repositories";
 import { getPickupArticles, getPopularArticles } from "@/services/article";
 import { getCategories } from "@/services/category";
 import type { PagePropsOrError } from "@/types";
